+++ android/core/ui/src/main/kotlin/com/andlife/ui/component/guestbook/GuestBookItem.kt
@@ -105,6 +105,8 @@ import com.andlife.model.guestbook.MediaUiType
 import com.andlife.ui.R
 import com.andlife.ui.component.icon.PlayerThumbnailIcon
 import com.andlife.ui.component.media.MediaOverlay
+import com.andlife.ui.component.media.video.FullscreenControlOverlay
+import com.andlife.ui.component.media.video.PlayerControlOverlay
 import com.andlife.ui.util.findActivity
 import com.andlife.ui.util.noRippleClickable
 import com.andlife.ui.util.toFormatDuration
@@ -743,159 +745,6 @@ private fun VideoPlayerContent(
     }
 }
 
-@Composable
-private fun PlayerControlOverlay(
-    progress: Float,
-    timeText: String,
-    isControlVisible: Boolean,
-    isMuted: Boolean,
-    isFullscreen: Boolean,
-    onSeekValueChange: (Float) -> Unit,
-    onSeekValueChangeFinished: () -> Unit,
-    onMuteToggle: () -> Unit,
-    onFullscreenClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Box(modifier = modifier.fillMaxSize()) {
-        AnimatedVisibility(
-            visible = !isControlVisible,
-            enter = fadeIn(),
-            exit = fadeOut(),
-            modifier = Modifier.align(Alignment.BottomCenter)
-        ) {
-            PlayerSeekbar(
-                progress = progress,
-                modifier = Modifier.fillMaxWidth()
-            )
-        }
-
-        AnimatedVisibility(
-            visible = isControlVisible,
-            enter = slideInVertically(
-                initialOffsetY = { it },
-                animationSpec = tween(durationMillis = 250)
-            ) + fadeIn(animationSpec = tween(durationMillis = 250)),
-            exit = slideOutVertically(
-                targetOffsetY = { it },
-                animationSpec = tween(durationMillis = 250)
-            ) + fadeOut(animationSpec = tween(durationMillis = 250)),
-            modifier = Modifier.align(Alignment.BottomCenter)
-        ) {
-            PlayerControlBar(
-                progress = progress,
-                timeText = timeText,
-                isFullscreen = isFullscreen,
-                onSeekValueChange = onSeekValueChange,
-                onSeekValueChangeFinished = onSeekValueChangeFinished,
-                onFullscreenClick = onFullscreenClick,
-            )
-        }
-
-        PlayerMuteButton(
-            isMuted = isMuted,
-            onToggle = onMuteToggle,
-            modifier = Modifier
-                .align(Alignment.TopStart)
-                .padding(NachoSpacing.small),
-        )
-    }
-}
-
-@kotlin.OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun PlayerControlBar(
-    progress: Float,
-    timeText: String,
-    isFullscreen: Boolean,
-    onSeekValueChange: (Float) -> Unit,
-    onSeekValueChangeFinished: () -> Unit,
-    onFullscreenClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-//    val fullScreenIconRes = if (isFullscreen) {
-//        R.drawable.ic_fullscreen_exit_24
-//    } else {
-//        R.drawable.ic_fullscreen_24
-//    }
-
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .background(
-                brush = Brush.verticalGradient(
-                    colors = listOf(Color.Transparent, Color.Black.copy(alpha = 0.6f))
-                )
-            )
-            .padding(horizontal = NachoSpacing.small)
-            .padding(bottom = NachoSpacing.medium),
-//        verticalArrangement = Arrangement.spacedBy(NachoSpacing.medium),
-    ) {
-        Slider(
-            value = progress,
-            onValueChange = onSeekValueChange,
-            onValueChangeFinished = onSeekValueChangeFinished,
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(24.dp),
-            thumb = {
-                Box(
-                    modifier = Modifier
-                        .size(NachoIconSize.xSmall)
-                        .shadow(
-                            elevation = NachoElevation.medium,
-                            shape = CircleShape
-                        )
-                        .background(
-                            color = Color.White,
-                            shape = CircleShape,
-                        )
-                )
-            },
-            track = { sliderState ->
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(NachoStroke.large)
-                        .clip(NachoTheme.shapes.extraSmall)
-                        .background(NachoTheme.colorScheme.backgroundBorder)
-                ) {
-                    Box(
-                        modifier = Modifier
-                            .fillMaxWidth(sliderState.value)
-                            .fillMaxHeight()
-                            .background(NachoTheme.colorScheme.brandPrimary)
-                    )
-                }
-            },
-        )
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = NachoSpacing.xSmall, vertical = NachoSpacing.xSmall),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Text(
-                text = timeText,
-                style = NachoTheme.typography.bodySmallRegular,
-                color = Color.White,
-                modifier = Modifier.padding(start = NachoSpacing.xSmall)
-            )
-            Icon(
-                painter = painterResource(R.drawable.ic_fullscreen_24),
-                contentDescription = stringResource(R.string.desc_fullscreen),
-                tint = Color.White,
-                modifier = Modifier
-                    .clickable(
-                        interactionSource = remember { MutableInteractionSource() },
-                        indication = null,
-                        onClick = onFullscreenClick
-                    ),
-            )
-        }
-    }
-}
-
 @SuppressLint("ConfigurationScreenWidthHeight")
 @OptIn(UnstableApi::class)
 @Composable
@@ -1100,187 +949,6 @@ fun VideoFullscreenOverlay(
     }
 }
 
-@Composable
-private fun FullscreenControlOverlay(
-    progress: Float,
-    timeText: String,
-    isControlVisible: Boolean,
-    isPlaying: Boolean,
-    isMuted: Boolean,
-    isLandscape: Boolean,
-    onPlayPauseClick: () -> Unit,
-    onSeekValueChange: (Float) -> Unit,
-    onSeekValueChangeFinished: () -> Unit,
-    onMuteToggle: () -> Unit,
-    onOrientationClick: () -> Unit,
-    onExitFullscreen: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Box(modifier = modifier.fillMaxSize()) {
-        AnimatedVisibility(
-            visible = isControlVisible,
-            enter = fadeIn(),
-            exit = fadeOut(),
-            modifier = Modifier.align(Alignment.Center),
-        ) {
-            Box(
-                modifier = Modifier
-                    .size(56.dp)
-                    .background(
-                        color = Color.Black.copy(alpha = 0.5f),
-                        shape = CircleShape,
-                    )
-                    .noRippleClickable { onPlayPauseClick() },
-                contentAlignment = Alignment.Center,
-            ) {
-                Icon(
-                    painter = painterResource(
-                        if (isPlaying) R.drawable.ic_pause_filled_24
-                        else R.drawable.ic_play_arrow_24
-                    ),
-                    contentDescription = if (isPlaying) "일시정지" else "재생",
-                    tint = Color.White,
-                    modifier = Modifier.size(32.dp),
-                )
-            }
-        }
-
-        AnimatedVisibility(
-            visible = isControlVisible,
-            enter = slideInVertically(
-                initialOffsetY = { it },
-                animationSpec = tween(durationMillis = 250),
-            ) + fadeIn(animationSpec = tween(durationMillis = 250)),
-            exit = slideOutVertically(
-                targetOffsetY = { it },
-                animationSpec = tween(durationMillis = 250),
-            ) + fadeOut(animationSpec = tween(durationMillis = 250)),
-            modifier = Modifier.align(Alignment.BottomCenter),
-        ) {
-            FullscreenControlBar(
-                progress = progress,
-                timeText = timeText,
-                isPlaying = isPlaying,
-                isMuted = isMuted,
-                isLandscape = isLandscape,
-                onPlayPauseClick = onPlayPauseClick,
-                onSeekValueChange = onSeekValueChange,
-                onSeekValueChangeFinished = onSeekValueChangeFinished,
-                onMuteToggle = onMuteToggle,
-                onOrientationClick = onOrientationClick,
-                onExitFullscreen = onExitFullscreen,
-            )
-        }
-    }
-}
-
-@kotlin.OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun FullscreenControlBar(
-    progress: Float,
-    timeText: String,
-    isPlaying: Boolean,
-    isMuted: Boolean,
-    isLandscape: Boolean,
-    onPlayPauseClick: () -> Unit,
-    onSeekValueChange: (Float) -> Unit,
-    onSeekValueChangeFinished: () -> Unit,
-    onMuteToggle: () -> Unit,
-    onOrientationClick: () -> Unit,
-    onExitFullscreen: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .background(
-                brush = Brush.verticalGradient(
-                    colors = listOf(Color.Transparent, Color.Black.copy(alpha = 0.6f))
-                )
-            )
-            .padding(horizontal = NachoSpacing.small)
-            .padding(bottom = NachoSpacing.xLarge),
-    ) {
-        Slider(
-            value = progress,
-            onValueChange = onSeekValueChange,
-            onValueChangeFinished = onSeekValueChangeFinished,
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(24.dp),
-            thumb = {
-                Box(
-                    modifier = Modifier
-                        .size(NachoIconSize.xSmall)
-                        .shadow(elevation = NachoElevation.medium, shape = CircleShape)
-                        .background(color = Color.White, shape = CircleShape)
-                )
-            },
-            track = { sliderState ->
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(NachoStroke.large)
-                        .clip(NachoTheme.shapes.extraSmall)
-                        .background(NachoTheme.colorScheme.backgroundBorder)
-                ) {
-                    Box(
-                        modifier = Modifier
-                            .fillMaxWidth(sliderState.value)
-                            .fillMaxHeight()
-                            .background(NachoTheme.colorScheme.brandPrimary)
-                    )
-                }
-            },
-        )
-
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = NachoSpacing.xSmall, vertical = NachoSpacing.xSmall),
-            horizontalArrangement = Arrangement.spacedBy(NachoSpacing.small),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Icon(
-                painter = painterResource(
-                    if (isPlaying) R.drawable.ic_pause_filled_24
-                    else R.drawable.ic_play_arrow_24
-                ),
-                contentDescription = if (isPlaying) "일시정지" else "재생",
-                tint = Color.White,
-                modifier = Modifier.noRippleClickable { onPlayPauseClick() },
-            )
-            Text(
-                text = timeText,
-                style = NachoTheme.typography.bodySmallRegular,
-                color = Color.White,
-            )
-            Spacer(modifier = Modifier.weight(1f))
-            Icon(
-                painter = painterResource(
-                    if (isMuted) R.drawable.ic_volume_off_filled_24
-                    else R.drawable.ic_volume_up_filled_24
-                ),
-                contentDescription = if (isMuted) "음소거 해제" else "음소거",
-                tint = Color.White,
-                modifier = Modifier.noRippleClickable { onMuteToggle() },
-            )
-            Icon(
-                painter = painterResource(R.drawable.ic_screen_rotation_24),
-                contentDescription = if (isLandscape) "세로 모드" else "가로 모드",
-                tint = Color.White,
-                modifier = Modifier.noRippleClickable { onOrientationClick() },
-            )
-            Icon(
-                painter = painterResource(R.drawable.ic_fullscreen_exit_24),
-                contentDescription = stringResource(R.string.desc_exit_fullscreen),
-                tint = Color.White,
-                modifier = Modifier.noRippleClickable { onExitFullscreen() },
-            )
-        }
-    }
-}
-
 @OptIn(UnstableApi::class)
 @Composable
 fun VideoPlayerView(
@@ -1313,73 +981,6 @@ private fun ThumbnailWrapper(
             contentScale = ContentScale.Fit,
         )
         PlayerThumbnailIcon(modifier = Modifier.align(Alignment.Center))
-    }
-}
-
-@Composable
-private fun VideoDurationOverlay(
-    duration: String,
-    modifier: Modifier = Modifier,
-) {
-    MediaOverlay(
-        modifier = modifier,
-        text = duration
-    )
-}
-
-@Composable
-private fun PlayerSeekbar(
-    progress: Float,
-    modifier: Modifier = Modifier,
-) {
-    LinearProgressIndicator(
-        progress = { progress },
-        modifier = modifier
-            .fillMaxWidth()
-            .height(NachoStroke.medium),
-        color = NachoTheme.colorScheme.brandPrimary,
-        trackColor = NachoTheme.colorScheme.backgroundBorder,
-        gapSize = 0.dp,
-        strokeCap = StrokeCap.Square,
-        drawStopIndicator = { /* No-op */ },
-    )
-}
-
-@Composable
-private fun PlayerMuteButton(
-    isMuted: Boolean,
-    onToggle: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val iconResId = if (isMuted) {
-        R.drawable.ic_volume_off_filled_24
-    } else {
-        R.drawable.ic_volume_up_filled_24
-    }
-    val contentDescription = if (isMuted) {
-        stringResource(R.string.desc_unmute_video)
-    } else {
-        stringResource(R.string.desc_mute_video)
-    }
-    Box(
-        modifier = modifier
-            .background(
-                color = NachoTheme.colorScheme.backgroundOverlay,
-                shape = CircleShape
-            )
-            .clickable(
-                interactionSource = remember { MutableInteractionSource() },
-                indication = null,
-                onClick = onToggle
-            ),
-        contentAlignment = Alignment.Center
-    ) {
-        Icon(
-            painter = painterResource(id = iconResId),
-            contentDescription = contentDescription,
-            tint = NachoTheme.colorScheme.iconTertiary,
-            modifier = Modifier.padding(NachoSpacing.xSmall)
-        )
     }
 }
